@michaelbromley 2021-02-11
Excellent find! I prepared a benchmark which:

* Creates an order
* Adds a random ProductVariant to it in a quantity of up to 500
* Repeats for 1 minute

Before:
```
http_req_duration..........: avg=2.26s   min=107.8ms med=1.76s   max=8.45s   p(90)=3.1s    p(95)=7.51s
http_reqs..................: 23      0.383328/s
```

After:
```
http_req_duration..........: avg=101.07ms min=86.99ms  med=97.97ms  max=328.97ms p(90)=107ms    p(95)=129ms
http_reqs..................: 592     9.866615/s
```

Incredible! A ~**25x speedup!**

Makes me wonder about what other easy perf wins are lurking in the code base... 🤔