@thomas-advantitge 2021-02-12
We just noticed the order items are not saved. The adjustment call returns the correct values, but executing a subsequent order call doesn't.

Unfortunately, this is due to a stupid mistake in the PR which causes the order items not to be saved due to the lack of `.execute()` 😞. It appears to be required for insert queries, while it isn't for select queries.

Luckily, after adding `.execute()`, the order item loop still performs well (it appears the `save` vs `insert` issue still was the bottleneck).

However, the second bottleneck, the `applyPriceAdjustments` section:
```typescript
await this.connection.getRepository(ctx, OrderItem).createQueryBuilder().insert().values(updatedItems).execute();
await this.connection.getRepository(ctx, ShippingLine).createQueryBuilder().insert().values(order.shippingLines).execute();
```
returns duplicate keys: we need an update query. This is not as easy as a bulk `.insert()`, actually this is exactly the goal of `.save()` , but this causes several seconds delay here too.

Two remarks:
* The shippingLines `.save()` delay seems to be neglectable.
* The best alternative I can come up with right now for the `updatedItems` `.save()` is an insert-update-on-duplicate-key query. It appears TypeORM has an `.orUpdate` function for this behaviour:
```typescript
await this.connection.getRepository(ctx, OrderItem).createQueryBuilder().insert().values(updatedItems).orUpdate({ conflict_target: ['id'], overwrite: ['taxLines', 'adjustments', 'listPriceIncludesTax'] }).execute();
```
I tested this quickly and it seems to be working, reducing ~3s calls to ~80ms, but it might need some extra testing. Also, explicitly listing the database fields which need to be overriden feels a bit ugly, especially as the complex calculation logic is hidden from the order service.